requirements.txt
SQLAlchemy~=2.0.40
PyMySQL==1.1.0
requests==2.31.0
python-dotenv~=1.0.0
pydantic~=2.11.3
pydantic-core==2.23.2
pydantic-settings~=2.9.1
fastapi==0.115.12
uvicorn~=0.34.2
sqlparse~=0.4.4
PyYAML~=6.0